Add nesting tests
We want to be able to test whether we can properly nest sandboxes.
Right now, we support two configurations: Unprivileged -> Unprivileged,
and Docker -> Unprivileged.  We do not support Privileged in any part of
the nesting chain, as it is only useful in cases where we know the
kernel doesn't support unprivileged user namespaces, which is exactly
what we need to run inside of the sandbox.