Test nested persistence
Because `overlayfs` mounts do not like to have their changes sit on top
of another overlay, we need to be careful when enabling persistence
mode.  When we are not persisting changes, we create a `tmpfs` and drop
the overlayfs changes onto that tmpfs.  However, when persisting we
naively assume that `/tmp` is a real directory and not itself an
overlay.  To work around this, we allow the parent directory to be
specified by a convenient environment variable,
`SANDBOX_PERSISTENCE_DIR`, which should be set to a read-write-mounted
directory on any sandbox that expects to perform nested sandboxing.