Sketch out initial multiarch support
This adds basic support for probing `binfmt_misc` presence, as well as
support for hooking the `qemu-*-static` binaries up to the kernel, in
the event that they are not already installed.  This should work
seamlessly for most installs, the rare edge cases being computers that
already have interpreters registered but do not have the `F` flag set
which causes the interpreter to be immediately loaded (And hence
available even in other mount namespaces), or non-functional
interpreters.

TODO: still need to write tests for each architecture.