Work around issue with persistent dirs with `UnprivilegedUserNamespacesExecutor`
Our efforts to solve this have been stymied by a number of confounding factors, but I think I may have finally wrangled it.  The issues include:

* On newer kernels, you need to set `userxattr` on overlay mounts, otherwise you get failures to `mv` directories (manifesting as `invalid cross-device link` within the sandbox, and `failed to set xattr on upper` in the kernel log). See https://www.kernel.org/doc/html/latest/filesystems/overlayfs.html#renaming-directories for a bit more detail.

* We cannot use `zfs` as the backing filesystem for an overlay's upper directory.  Trying to do this will manifest in `mount()` failing within the sandbox, and the kernel log complaining about a lack of the `RENAME_WHITEOUT` feature.

* Ubuntu patches spoil lazy devs like me.

* Many of these filesystem rules work one way in the host environment, and a different way when inside of the sandbox, which means that nested sandboxing is often the most difficult challenge.

I've made some changes here to default to `persist = true`.  Not only is it a cool and useful feature with very little overhead, it makes it a lot easier on `userxattr` systems.  We have a fancy new search algorithm for a directory that is able to be mounted against, which should aid us in nested sandboxing.

Because we're changing the `persist` default here, I think this is worth bumping version numbers over.  There's also a significant amount of code here, so I may want to take a second pass in a few days to try and re-organize things a bit.