Add signal forwarding/reflection
We want to be able to send things like  `SIGTERM` to the process we
started in the sandbox, and so we forward many useful signals through.
If our child then dies with a signal, we pass this information back up
to the parent by communicating through pipes to dodge the issue of PID 1
within a user namespace not processing signals properly.