Improve sandbox `run()` wrapper to support `wait=false`.
This solution is not totally satisfactory, as it seems a little race-y.
It would be better if we somehow were able to return the `Task` given by
the `@async`, but if we do that, we can't do things like send signals to
the process we just launched, because we're no longer returning a
`Base.Process`, we're returning a `Task`.