Switch over to using multiarch'ed rootfs images for probing
This drops many of the rootfs images that were included, choosing
instead to consolidate on a minimal debian rootfs image.  Users can
continue to use the Debian rootfs for testing (especially as it will now
correctly choose a rootfs for the current architecture) but an even
better plan is to use custom rootfs images.