Overlayed mount types
This commit performs a reasonably large overhaul of the mounting logic
in the sandbox, providing support for "overlay" mounts.  This support
allows the user to specify that certain mountpoints should be covered by
an overlayfs, allowing modifications to be stored in the persistence dir
rather than in the mountpoint itself.

This is a major change, and so with it we get a new major release!