Fix race condition in sandbox
The current sandbox code has a race condition. If the child exits
before the sigwait signal mask is setup, we will hang forever waiting
for a signal. This has been observed in practice in Yggdrasil.
I don't know that the sigwait here is supposed to do - plain
waitpid should be fine, so try removing it to see if everything
works.

If the sigwait does turn out to be necessary, we should be able
to fix this by moving the signal mask setup to above the fork instead.